mag_bot/requirements.txt
python-telegram-bot==20.3
selenium
webdriver-manager
pandas
numpy
scipy
matplotlib
